# Conflicts:
#	build.gradle.kts
#	docker-compose.yml
#	src/main/java/com/codeofus/rent_a_park/configuration/RedisConfiguration.java
#	src/main/java/com/codeofus/rent_a_park/controllers/PersonController.java
#	src/main/java/com/codeofus/rent_a_park/dtos/ParkingMapper.java
#	src/main/java/com/codeofus/rent_a_park/dtos/PersonDto.java
#	src/main/java/com/codeofus/rent_a_park/errors/BadRequestException.java
#	src/main/java/com/codeofus/rent_a_park/models/Person.java
#	src/main/java/com/codeofus/rent_a_park/models/Spot.java
#	src/main/java/com/codeofus/rent_a_park/services/PersonService.java
#	src/test/java/com/codeofus/rent_a_park/IntegrationTest.java
#	src/test/java/com/codeofus/rent_a_park/PersonControllerTests.java
#	src/test/java/com/codeofus/rent_a_park/SpotControllerTests.java

+++ src/main/java/com/codeofus/rent_a_park/controllers/ReservationController.java
@@ -30,7 +30,7 @@ public class ReservationController {
     }
 
     @GetMapping("/{id}")
-    public ReservationDto getReservation(@PathVariable long id) throws BadRequestException {
+    public ReservationDto getReservation(@PathVariable int id) throws BadRequestException {
         Optional<Reservation> reservation = reservationService.getReservation(id);
         if (reservation.isPresent()) {
             return reservationMapper.reservationToReservationDTO(reservation.get());
@@ -59,7 +59,7 @@ public class ReservationController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteReservation(@PathVariable Long id) {
+    public void deleteReservation(@PathVariable int id) {
         reservationService.deleteReservation(id);
     }
 }
